requirements.txt
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.3.0
tensorflow>=2.12.0
joblib>=1.3.0
matplotlib>=3.5.0
streamlit>=1.28.0
plotly>=5.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
requests>=2.28.0
# Authentication dependencies
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
sqlalchemy>=2.0.0
streamlit-authenticator>=0.2.3
streamlit-option-menu>=0.3.6
